requirements/base.txt
adrf
celery
Django
django-cors-headers
django-debug-toolbar
django-extensions
django-filter
django-import-export[all]
django-jazzmin
django-ratelimit
django-redis
django-schema-viewer
djangorestframework
djangorestframework-jsonapi[django-filter]
djangorestframework-simplejwt
drf-material
drf-spectacular
drf-spectacular[sidecar]
drf-standardized-errors
Faker
flower
icecream
phonenumbers
pillow
psycopg2-binary
python-decouple
whitenoise
